CSE 360 Assignment 3 - Third Version
Third Version Requirements
a. Make the changes to the Calculator class as needed so that the getHistory method
will return a history of all actions as a String.
  i. The history should start from the initial 0 value to the last input.
  ii. Use the normal operators for the operations (+, -, *, /). There is no limit
on the size of the String returned.
  iii. The return String should separate all operations and values with a space.
    For example, if the following actions are done,
    myCalculator.add (4);
    myCalculator.subtract (2);
    myCalculator.multiply (2);
    myCalculator.add(5);
    then the value returned should be
    0 + 4 – 2 * 2 + 5
  iv. Add other variables as needed to the class. Make changes to the other
  methods as needed.
b. Update the CalculatorTest to test the getHistory method.
c. Submit both java files to git and github